JME functions to query the board; resume; jsxgraph 1.3.2
The `jsxgraph` and `jessiecode` functions now return an object of type
`jsxgraphboard`.

It can be accessed like a dictionary to get objects by id.

There are a load of functions of the form `jxg_property` to get
properties of objects, which can be used in marking algorithms.

`jsxgraphboard` objects are correctly saved and restored from suspend
data. It's important that they don't use any custom JS, because the
JessieCode dump can't deal with that.

This commit also updates JSXGraph to version 1.3.2