lots more changes
The link to parts is now more robust, and copes with several parts in a
question being linked to boards.

When there's an error in the jxg_input or jxg_output notes, it's
shown as a part warning.

Objects created through JME can supply a dictionary of event hooks, in
the form of a list [event name, jessiecode to run]

jxg_has_point now converts to screen coords properly, and can take a
vector instead of two numbers.

The functions modfifying the board produce 'tjsxgraphaction' tokens,
which are only applied by the bind_input function. This means that when
the note is evaluated by the normal marking algorithm, it doesn't lead
to a board update.

There are jxg_add_objects and jxg_run_jessiecode functions which can be
used to add to existing boards. I defined a function which creates a
generic board, and then I use these to add in different things for each
instance.