correctly resume boards that weren't drawn
This changes the behaviour of TJSXGraphBoard objects when the board has
not been created yet. Any JessieCode scripts or lists of JME-defined
objects are saved on the TJSXGraphBoard value, and if the board has not
been drawn yet then they are serialised as an argument to a new function
`resume_jsxgraph_board_init`.

This means that the board can be recreated when the attempt is resumed.

fixes #1